Changed how simple injector starts its scope.
Resolving a dependency is as follows. Firstly, SI looks to see if it's in AsyncScope, if it is not, then secondly it checks if it is in a WebRequestScope. WebApi will respect async flow and thus resolve in the AsyncScope, MVC will destroy the async flow and thus resolve in the WebRequestScope. SignalR will also destroy the async flow for (maybe?) the same reasons. However, it cannot resolve in a WebRequestScope as it does not treat hub requests as web requests (HttpContext is null) or even as Owin requests (the owin pipeline is not triggered) and thus will need special implementation to fix.