Added tests for the ResetSecurityStamp request handler and the ValidationPipelineBehavior pipeline.
Fixed the in memory data context (which should just be replaced with one that actually supports async operations like a normal database context, but that's for another MR later and possibly another rant as well >_>) ... so it supports async operations.

Added a builder for the new entity that splits off of the User table.